+++ sdk/data/azcosmos/example_test.go
@@ -8,10 +8,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos"
 )
 
@@ -64,7 +64,9 @@ func ExampleClient_CreateDatabase() {
 	databaseProperties := azcosmos.DatabaseProperties{ID: "databaseName"}
 	databaseResponse, err := client.CreateDatabase(context.Background(), databaseProperties, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Database created. ActivityId %s", databaseResponse.ActivityID)
@@ -107,7 +109,9 @@ func ExampleDatabaseClient_CreateContainer() {
 
 	resp, err := database.CreateContainer(context.Background(), properties, &azcosmos.CreateContainerOptions{ThroughputProperties: &throughput})
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Container created. ActivityId %s", resp.ActivityID)
@@ -154,7 +158,9 @@ func ExampleContainerClient_ReplaceThroughput() {
 	newScale := azcosmos.NewManualThroughputProperties(500)
 	replaceThroughputResponse, err := container.ReplaceThroughput(context.Background(), newScale, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Throughput updated. ActivityId %s", replaceThroughputResponse.ActivityID)
@@ -202,7 +208,9 @@ func ExampleContainerClient_Replace() {
 	// Replace container properties
 	replaceResponse, err := container.Replace(context.Background(), *containerResponse.ContainerProperties, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Container updated. ActivityId %s", replaceResponse.ActivityID)
@@ -249,7 +257,9 @@ func ExampleContainerClient_CreateItem() {
 
 	itemResponse, err := container.CreateItem(context.Background(), pk, marshalled, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Item created. ActivityId %s consuming %v RU", itemResponse.ActivityID, itemResponse.RequestCharge)
@@ -286,7 +296,9 @@ func ExampleContainerClient_ReadItem() {
 	id := "anId"
 	itemResponse, err := container.ReadItem(context.Background(), pk, id, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	var itemResponseBody map[string]string
@@ -347,7 +359,9 @@ func ExampleContainerClient_ReplaceItem() {
 
 	itemResponse, err = container.ReplaceItem(context.Background(), pk, id, marshalledReplace, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Item replaced. ActivityId %s consuming %v RU", itemResponse.ActivityID, itemResponse.RequestCharge)
@@ -384,7 +398,9 @@ func ExampleContainerClient_DeleteItem() {
 	id := "anId"
 	itemResponse, err := container.DeleteItem(context.Background(), pk, id, nil)
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Item deleted. ActivityId %s consuming %v RU", itemResponse.ActivityID, itemResponse.RequestCharge)
@@ -440,7 +456,9 @@ func ExampleContainerClient_ReadItem_sessionConsistency() {
 	// In another client, maintain the session by passing the session token
 	itemResponse, err = container.ReadItem(context.Background(), pk, id, &azcosmos.ItemOptions{SessionToken: itemSessionToken})
 	if err != nil {
-		panic(err)
+		var responseErr *azcore.ResponseError
+		errors.As(err, &responseErr)
+		panic(responseErr)
 	}
 
 	fmt.Printf("Item read. ActivityId %s consuming %v RU", itemResponse.ActivityID, itemResponse.RequestCharge)
@@ -501,12 +519,61 @@ func ExampleContainerClient_ReplaceItem_optimisticConcurrency() {
 		// Replace with Etag
 		etag := itemResponse.ETag
 		itemResponse, err = container.ReplaceItem(context.Background(), pk, id, marshalledReplace, &azcosmos.ItemOptions{IfMatchEtag: &etag})
-		var httpErr interface{ RawResponse() *http.Response }
+		var responseErr *azcore.ResponseError
 
-		return (errors.As(err, &httpErr) && itemResponse.RawResponse.StatusCode == 412), err
+		return (errors.As(err, &responseErr) && responseErr.StatusCode == 412), err
 	})
 	if err != nil {
 		panic(err)
+	}
+}
+
+func ExampleContainerClient_NewQueryItemsPager() {
+	endpoint, ok := os.LookupEnv("AZURE_COSMOS_ENDPOINT")
+	if !ok {
+		panic("AZURE_COSMOS_ENDPOINT could not be found")
+	}
+
+	key, ok := os.LookupEnv("AZURE_COSMOS_KEY")
+	if !ok {
+		panic("AZURE_COSMOS_KEY could not be found")
+	}
+
+	cred, err := azcosmos.NewKeyCredential(key)
+	if err != nil {
+		panic(err)
+	}
+
+	client, err := azcosmos.NewClientWithKey(endpoint, cred, nil)
+	if err != nil {
+		panic(err)
+	}
+
+	container, err := client.NewContainer("databaseName", "aContainer")
+	if err != nil {
+		panic(err)
+	}
+
+	pk := azcosmos.NewPartitionKeyString("newPartitionKey")
+
+	queryPager := container.NewQueryItemsPager("select * from docs c", pk, nil)
+	for queryPager.More() {
+		queryResponse, err := queryPager.NextPage(context.Background())
+		if err != nil {
+			var responseErr *azcore.ResponseError
+			errors.As(err, &responseErr)
+			panic(responseErr)
+		}
+
+		for _, item := range queryResponse.Items {
+			var itemResponseBody map[string]interface{}
+			err = json.Unmarshal(item, &itemResponseBody)
+			if err != nil {
+				panic(err)
+			}
+		}
+
+		fmt.Printf("Query page received with %v items. ActivityId %s consuming %v RU", len(queryResponse.Items), queryResponse.ActivityID, queryResponse.RequestCharge)
 	}
 }